mmwave_60g/requirement.txt
astroid==2.11.5
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fail2ban==0.11.1
httplib2==0.14.0
idna==2.8
isort==5.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
louis==3.12.0
macaroonbakery==1.3.1
mccabe==0.7.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==3.6.1
py4j==0.10.9.3
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.13.9
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyserial==3.5
pyspark==3.2.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
tomli==2.0.1
typing-extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wrapt==1.14.1
xkit==0.0.0
